Text Transformer updates (#43)
* Move to correct folder, strip example to be as simple as possible

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci
